finai-analyst/requirements.txt
streamlit
PyMuPDF
openai
google-generativeai
python-dotenv
chromadb
langchain
langchain-community
langchain-openai
tiktoken